修改抢红包
Summary: 修改抢红包

Test Plan: local test

Reviewers: chenqixiang, yingming.zhao, guobin.chu, lin.huang, lihaifeng

Reviewed By: chenqixiang

Differential Revision: https://pha.static.kunlun.com/D17131

+++ im/src/main/java/com/opay/im/model/response/LuckyMoneyResponse.java
@@ -52,7 +52,7 @@ public class LuckyMoneyResponse {
      * 目标类型:0个人 1群
      */
     @ApiModelProperty(value="目标类型:0个人 1群")
-    private Byte targetType;
+    private Integer targetType;
 
     /**
      * 红包说明Best Wishes!
